Refactoring Pages
