+++ components/LoginComponent/styles.js
@@ -1,5 +1,6 @@
 // ------------------------------ import libraries
 import styled from "styled-components";
+import Link from "next/link";
 // import { media } from "../../global-styles/MediaQueries";
 
 // // ------------------------------ import variables
@@ -25,7 +26,7 @@ export const LoginBox = styled.div`
   align-items: center;
 `;
 
-export const LoginButton = styled.button`
+export const LoginButton = styled(Link)`
   height: 40px;
   width: 100px;
   border-radius: 10px;
