+++ pages/index.js
@@ -1,44 +1,28 @@
 import React from "react";
-import fetch from "isomorphic-unfetch";
+import { LoginComponent } from "../components/LoginComponent";
+// import Image from "next/image";
 
-import { SearchComponent } from "../components/SearchComponent";
-// import { LoginComponent } from "../components/LoginComponent";
 import {
-  IndexContainer,
-  IndexHeader,
-  IndexLogo,
-  Favorites,
-} from "../components/PagesStyles/indexStyles";
-import { MainComponent } from "../components/MainComponent";
+  LoginContainer,
+  LoginLogo,
+} from "../components/PagesStyles/loginStyles";
 import { GlobalStyles } from "../components/PagesStyles/GlobalStyles";
 
-const Home = (props) => {
+// import Image from "../../assets/image";
+// const dev = require("/dev.jpg");
+
+const LoginPage = () => {
   return (
     <>
       <GlobalStyles />
-      <IndexContainer>
-        <IndexHeader>
-          <IndexLogo>
-            <img src="/image/logo2.png" alt="Platzi Logo" />
-          </IndexLogo>
-          <h1>Talent Placement</h1>
-          {console.log(props.offers)}
-          <Favorites>Favoritos</Favorites>
-        </IndexHeader>
-        <SearchComponent />
-        <MainComponent offers={props.offers} />
-      </IndexContainer>
+      <LoginContainer>
+        <LoginLogo>
+          <img src="/image/logo2.png" alt="Platzi Logo" />
+        </LoginLogo>
+        <LoginComponent />
+      </LoginContainer>
     </>
   );
 };
 
-Home.getInitialProps = async (ctx) => {
-  const res = await fetch("https://hackaton-master.herokuapp.com/jobs");
-  const resJSON = await res.json();
-  return {
-    offers: resJSON.data,
-  };
-};
-// console.log(props.users);
-
-export default Home;
+export default LoginPage;
